pyproject.toml
[project]
name = "agui-test-frameworks"
version = "0.1.0"
description = "AG-UI Multi-Framework Test Suite"
requires-python = ">=3.10"
dependencies = [
    # Common
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",

    # Agno Framework
    "agno>=1.0.0",

    # LangGraph Framework
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.2.0",
    "ag-ui-langgraph>=0.0.20",

    # CrewAI Framework
    "crewai>=0.80.0",
    "ag-ui-crewai>=0.0.1",

    # PydanticAI Framework
    "pydantic-ai>=0.1.0",

    # LlamaIndex Framework
    "llama-index-protocols-ag-ui>=0.2.0",
    "llama-index-core>=0.10.0",
    "llama-index-llms-openai>=0.1.0",
    "llama-index-llms-anthropic>=0.6.0",
    "llama-index-llms-gemini>=0.4.0",

    # AG2 (AutoGen) Framework
    "ag2[openai]>=0.6.0",

    # Google ADK Framework
    "google-adk>=1.0.0",

    # Raw LLM APIs
    "openai>=1.0.0",
    "google-generativeai>=0.8.0",
]

[tool.uv]
dev-dependencies = []

[tool.uv.workspace]
members = [
    "a2a-playground",
]
